refactor: BuyTogether 본인 Post 삭제 개선
SQL로 모든걸 처리하려 하지 않음
- 최대한 로직으로 처리하여 보다 더 직관적으로 보이게 함
- 다양한 에러를 나타낼수 있게 하였음 (본인 post가 아니면 못지웁니다 에러)

+++ src/main/java/kr/anabada/anabadaserver/domain/save/service/BuyTogetherService.java
@@ -55,8 +55,11 @@ public class BuyTogetherService {
 
     @Transactional
     public void removeMyPost(User user, Long buyTogetherId) {
-        BuyTogether buyTogether = saveRepository.findBuyTogetherByIdAndWriter(buyTogetherId, user)
+        BuyTogether buyTogether = saveRepository.findBuyTogetherById(buyTogetherId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 게시물이 없습니다."));
+
+        if (!buyTogether.getWriter().getId().equals(user.getId()))
+            throw new IllegalArgumentException("해당 게시물의 작성자가 아닙니다.");
 
         saveRepository.delete(buyTogether);
     }
